requirements.txt
# ── Core Web Framework ─────────────────────────────────
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.0
python-multipart>=0.0.9
aiofiles>=24.1.0

# ── Database — SQLAlchemy Async ORM ───────────────────
sqlalchemy[asyncio]>=2.0.0
alembic>=1.13.0
aiosqlite>=0.20.0          # SQLite async driver (dev/fallback)
psycopg2-binary>=2.9.9     # PostgreSQL driver (production)
asyncpg>=0.29.0            # PostgreSQL async driver for SQLAlchemy

# ── Auth — JWT & Sessions ──────────────────────────────
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# ── Cache & Task Queue ─────────────────────────────────
redis>=5.0.0
celery>=5.3.0
flower>=2.0.0              # Celery monitoring UI

# ── Monitoring ─────────────────────────────────────────
prometheus-client>=0.20.0

# ── Media Processing ───────────────────────────────────
opencv-python-headless>=4.10.0
Pillow>=10.0.0

# ── Network & Streaming ────────────────────────────────
websockets>=12.0
mss>=9.0.0
sounddevice>=0.4.6
numpy>=1.26.0

# ── Service Discovery & QR ─────────────────────────────
zeroconf>=0.130.0
qrcode[pil]>=7.4

# ── ASR (Optional — large download) ───────────────────
faster-whisper>=1.0.0

# ── Testing ────────────────────────────────────────────
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0              # AsyncClient for FastAPI TestClient
anyio>=4.0.0
